Remove self-dependency on stomp
Currently installing 1.4.2 through 'gem install stomp' or 'bundle install' breaks.
The reason is just that stomp relies on itself on the gemspec, the error
we get is:

"Your bundle requires gems that depend on each other, creating an
infinite loop. Please remove gem 'stomp' and try again."

This change simply removes that dependency that I believe should not be
there, and also removes the double dev dependency on rspec. Rake is
already required in the gemspec so I removed it from the Gemfile