Fix subtle poor behavior of suppress_content_length:
The suppress_content_length message header has been
supported since Wed Dec 23 16:23:03 2009 -0200.

Since then the implementation has been wasteful with
unneeded and unrequired functionality.

Gem behavior has been as follows.

Suppose the client:

-- Asks for suppress_content_length and
-- Supplied a message body of "abc\00def"

The gem correctly did not add a content-length header.
However, the entire seven bytes (in the above example)
was written to the TCP connection's buffer.

This can cause errors to be generated by some brokers,
and the last three bytes (in this example) do not
need to be written (suppose it were 3M bytes).

This behavior was uncovered during initial testing with
Artemis.

Continuing with that example, this commit changes gem
behavior to:
- Only write "abc\00" to the connection buffer
- Ignore the bytes "def"